Change which event blocks were created in (#535)
