+++ src/main/java/gregtech/common/CommonProxy.java
@@ -12,9 +12,12 @@ import gregtech.api.recipes.recipeproperties.FusionEUToStartProperty;
 import gregtech.api.recipes.recipeproperties.TemperatureProperty;
 import gregtech.api.terminal.TerminalRegistry;
 import gregtech.api.unification.material.Material;
+import gregtech.api.unification.material.Materials;
 import gregtech.api.unification.material.properties.DustProperty;
+import gregtech.api.unification.material.properties.FluidPipeProperties;
 import gregtech.api.unification.material.properties.PropertyKey;
 import gregtech.api.unification.ore.OrePrefix;
+import gregtech.api.unification.ore.StoneType;
 import gregtech.api.util.GTLog;
 import gregtech.api.util.advancement.GTTrigger;
 import gregtech.common.advancement.GTTriggers;
@@ -53,9 +56,12 @@ import net.minecraftforge.fml.common.eventhandler.EventPriority;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.registries.IForgeRegistry;
+import org.apache.commons.lang3.ArrayUtils;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
+import java.util.Arrays;
+import java.util.HashMap;
 import java.util.function.Function;
 
 import static gregtech.common.blocks.MetaBlocks.*;
@@ -69,6 +75,43 @@ public class CommonProxy {
         IForgeRegistry<Block> registry = event.getRegistry();
 
         registry.register(MACHINE);
+
+        StoneType.init();
+
+        for (Material material : GregTechAPI.MATERIAL_REGISTRY) {
+
+           if (material.hasProperty(PropertyKey.ORE)) {
+                createOreBlock(material);
+            }
+
+            if (material.hasProperty(PropertyKey.WIRE)) {
+                for (BlockCable cable : CABLES) {
+                    if (!cable.getItemPipeType(null).isCable() || !material.getProperty(PropertyKey.WIRE).isSuperconductor())
+                        cable.addCableMaterial(material, material.getProperty(PropertyKey.WIRE));
+                }
+            }
+            if (material.hasProperty(PropertyKey.FLUID_PIPE)) {
+                for (BlockFluidPipe pipe : FLUID_PIPES) {
+                    if(!pipe.getItemPipeType(pipe.getItem(material)).getOrePrefix().isIgnored(material)) {
+                        pipe.addPipeMaterial(material, material.getProperty(PropertyKey.FLUID_PIPE));
+                    }
+                }
+            }
+            if (material.hasProperty(PropertyKey.ITEM_PIPE)) {
+                for (BlockItemPipe pipe : ITEM_PIPES) {
+                    if(!pipe.getItemPipeType(pipe.getItem(material)).getOrePrefix().isIgnored(material)) {
+                        pipe.addPipeMaterial(material, material.getProperty(PropertyKey.ITEM_PIPE));
+                    }
+                }
+            }
+        }
+        for (BlockFluidPipe pipe : FLUID_PIPES) {
+            if(!pipe.getItemPipeType(pipe.getItem(Materials.Wood)).getOrePrefix().isIgnored(Materials.Wood) ||
+                    !pipe.getItemPipeType(pipe.getItem(Materials.TreatedWood)).getOrePrefix().isIgnored(Materials.TreatedWood)) {
+                pipe.addPipeMaterial(Materials.Wood, new FluidPipeProperties(310, 5, false));
+                pipe.addPipeMaterial(Materials.TreatedWood, new FluidPipeProperties(310, 8, false));
+            }
+        }
 
         for (BlockCable cable : CABLES) registry.register(cable);
         for (BlockFluidPipe pipe : FLUID_PIPES) registry.register(pipe);
@@ -114,6 +157,35 @@ public class CommonProxy {
         FRAMES.values().stream().distinct().forEach(registry::register);
         SURFACE_ROCK.values().stream().distinct().forEach(registry::register);
         ORES.forEach(registry::register);
+    }
+
+    private static void createOreBlock(Material material) {
+        StoneType[] stoneTypeBuffer = new StoneType[16];
+        int generationIndex = 0;
+        for (StoneType stoneType : StoneType.STONE_TYPE_REGISTRY) {
+            int id = StoneType.STONE_TYPE_REGISTRY.getIDForObject(stoneType), index = id / 16;
+            if (index > generationIndex) {
+                createOreBlock(material, copyNotNull(stoneTypeBuffer), generationIndex);
+                Arrays.fill(stoneTypeBuffer, null);
+            }
+            stoneTypeBuffer[id % 16] = stoneType;
+            generationIndex = index;
+        }
+        createOreBlock(material, copyNotNull(stoneTypeBuffer), generationIndex);
+    }
+
+    private static <T> T[] copyNotNull(T[] src) {
+        int nullIndex = ArrayUtils.indexOf(src, null);
+        return Arrays.copyOfRange(src, 0, nullIndex == -1 ? src.length : nullIndex);
+    }
+
+    private static void createOreBlock(Material material, StoneType[] stoneTypes, int index) {
+        BlockOre block = new BlockOre(material, stoneTypes);
+        block.setRegistryName("ore_" + material + "_" + index);
+        for (StoneType stoneType : stoneTypes) {
+            GregTechAPI.oreBlockTable.computeIfAbsent(material, m -> new HashMap<>()).put(stoneType, block);
+        }
+        ORES.add(block);
     }
 
     @SubscribeEvent(priority = EventPriority.LOWEST)
